Super Echo
Former-commit-id: 392432cb7c10cae785f0d60abf08b4ad23728161
Former-commit-id: 23e3b88d2743620eb4943cfa6dcd349445757b1e

+++ src/general_cmds.py
@@ -1,7 +1,6 @@
 from utils.utilities import *
 from utils.mongo_interface import *
 
-import discord
 import aiohttp
 
 class General(commands.Cog):
@@ -25,20 +24,40 @@ class General(commands.Cog):
 
     @commands.command()
     @commands.check(isnotThreat)
-    async def echo(self, ctx:commands.Context, text:str, tgt=None):
+    async def echo(self, ctx:commands.Context, text:str, tgt=None, deleteMsg=True):
         '''
         Echoes a block of text as if it were sent by someone else.
         Defaults to the author of the message is no target is given.
         Can mention by nickname.
         '''
 
-        if not tgt: tgt = ctx.author
-        else: tgt = ctx.message.mentions[0]
+        if not tgt: 
+            tgt = ctx.author
+        elif getUserfromNick(tgt):
+            tgt = ctx.guild.get_member((getUserfromNick(tgt))["_id"])
+        else: 
+            tgt = ctx.message.mentions[0]
 
         webhook = discord.Webhook.partial(707756204969033731, 'fBGEs9CYku6Cs9qbXTfu1HkyycZnR9zQbCbdlL7o3b0ul7xbHArae7B0pBCRMsBfX3Wy', adapter=discord.RequestsWebhookAdapter())
         #webhook.edit()
         webhook.send(text, username=tgt.display_name, avatar_url=tgt.avatar_url)
-        await ctx.message.delete()
+        if deleteMsg: await ctx.message.delete()
+    
+    @commands.command()
+    @commands.check(isnotThreat)
+    async def everyonesays(self, ctx:commands.Context, text:str):
+        '''
+        Please don't use this oh god
+        '''
+        count = 5
+
+        mems = list(ctx.guild.members)
+        random.shuffle(mems)
+
+        for member in mems:
+            await self.echo(ctx, text, member.display_name, False)
+            count -= 1
+            if count < 0: break
 
     @commands.command()
     async def buymefood(self, ctx:commands.Context):
@@ -51,6 +70,8 @@ class General(commands.Cog):
     async def avatar(self, ctx, nickname):
         '''
         Displays the avatar of the said person
+
+        Made by Slyflare and PhtephenLuu
         '''
         if not (getUserfromNick(nickname)):
             await ctx.send("Member with username " + nickname + " not found.")
@@ -64,6 +85,8 @@ class General(commands.Cog):
     async def test_userinfo(self, ctx, member = None):
         '''
         testestestesstetstststs
+
+        Made by Slyflare
         '''
         member = ctx.author if not member else ctx.message.mentions[0]
         roles = [role for role in member.roles]
